requirements.txt
telethon==1.34.0
google-auth==2.28.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1